+++ StoneOfAdventure_2019_UnityProject/Assets/Artifacts/HealthRegen/HealthRegen_artifact.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class HealthRegen_artifact : Artifact
 {
-    [SerializeField] private float healValue;
+    [SerializeField] private int healValue;
     Health Health;
 
     public void AddRegen()
